Update main.py
add covid data function from api.py and set as commend mean when it click the fetch data button it call the function

what is different b/w covid_data and covid_data()?

if we set commend = covid-data() it show data with out clicking the button b/c it call the function directly

when we use commend = covid_data it point the function when we cilck the button than it call the function ok.